Agrego Correciones de BD y demas.

+++ Domain/Entidades/DocenteAsignatura.cs
@@ -7,12 +7,13 @@ namespace Domain.Entidades
 {
     public class DocenteAsignatura: BaseEntity
     {
-        public int DocenteId { get; private set; }
+        public long DocenteId { get; private set; }
         public int AsignaturaId { get; private set; }
         public Docente Docente { get; private set; }
         public Asignatura Asignatura { get; private set; }
 
         public DocenteAsignatura() { }
+
         public DocenteAsignatura(Docente docente, Asignatura asignatura)
         {
             Docente = docente;
