+++ Domain/Entidades/DocenteCurso.cs
@@ -6,12 +6,13 @@ namespace Domain.Entidades
 {
     public class DocenteCurso
     {
-        public int DocenteId { get; private set; }
+        public long DocenteId { get; private set; }
         public int CursoId { get; private set; }
         public Docente Docente { get; private set; }
         public Curso Curso { get; private set; }
 
         public DocenteCurso() { }
+
         public DocenteCurso(Docente docente, Curso curso)
         {
             Docente = docente;
